requirements.test.txt
pytest
# Pinned to 3.0.0 due to dependency conflicts in CI/CD test environments
pytest-cov==3.0.0
# Pinned to 0.20.3 to fix test event loop and dependency resolution issues in GitHub Actions
pytest-asyncio==0.20.3
pytest-homeassistant-custom-component
hoymiles-wifi>=0.2.1
